Namespace all the services for cleaner class names
