+++ app/routes/products.rb
@@ -2,8 +2,10 @@ module Shop
   module Routes
     class Products < Base
       get '/products' do
-        products = Services::FetchProducts.new.call
-        products_presenters = products.map { |product| Presenters::Product.new(product) }
+        products = Services::Products::FetchAll.new.call
+        products_presenters = products.map do |product|
+          Presenters::Product.new(product)
+        end
         erb :'products/index', locals: { products: products_presenters }
       end
     end
