+++ app/services/cart_items/fetch_all.rb
@@ -1,0 +1,11 @@
+module Shop
+  module Services
+    module CartItems
+      class FetchAll
+        def call
+          CART_ITEMS
+        end
+      end
+    end
+  end
+end
