--- app/services/fetch_cart_items.rb
@@ -1,9 +1,0 @@
-module Shop
-  module Services
-    class FetchCartItems
-      def call
-        CART_ITEMS
-      end
-    end
-  end
-end
