--- app/services/fetch_product.rb
@@ -1,9 +1,0 @@
-module Shop
-  module Services
-    class FetchProduct
-      def call(product_id:)
-        PRODUCTS.find { |product| product.id == product_id }
-      end
-    end
-  end
-end
