--- app/services/fetch_products.rb
@@ -1,9 +1,0 @@
-module Shop
-  module Services
-    class FetchProducts
-      def call
-        PRODUCTS
-      end
-    end
-  end
-end
