+++ app/services/products/fetch_all.rb
@@ -1,0 +1,11 @@
+module Shop
+  module Services
+    module Products
+      class FetchAll
+        def call
+          PRODUCTS
+        end
+      end
+    end
+  end
+end
