--- app/services/remove_product_from_cart.rb
@@ -1,9 +1,0 @@
-module Shop
-  module Services
-    class RemoveProductFromCart
-      def call(product_id:)
-        CART_ITEMS.delete_if { |cart_item| cart_item.product.id == product_id }
-      end
-    end
-  end
-end
